Big Changes (Replicator + AirportDisplay)
I used the replicator to make copies of the main placeholder system node instead of having 5 unlinked systems merged together in the end (which meant that the changes in one placeholder weren't applied to the others. With the replicator now everything is centralized in one node). This means that we can create more copies of the placeholder instead of the initial 6.

Also added the Airport display option, just for fun. In order to show more letters on the display.